Updated and fixes test runner, go paths and restructure to work
