+++ cloudflarewarp_test.go
@@ -6,7 +6,7 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	plugin "github.com/BetterCorp/cloudflarewarp/src/cloudflarewarp"
+	plugin "github.com/BetterCorp/cloudflarewarp"
 )
 
 func TestNew(t *testing.T) {
